Build 4447: Fix errors building OpenVPN git-master version
Adds a patch to the OpenVPN git-master copy so it includes config.guess
and config.sub taken from the most recent OpenVPN release, 2.3.8.

This fixes two errors when building (although the errors can be
ignored, they are unnerving). On Xcode 7, a third error is generated
because of the two errors, so that error is also avoided.