Build 4465: Enhance preferences
Adds “primary” forced preferences override Deployed forced preferences
which override the user’s normal preferences.

To set them when TB is being installed, include
“forced-preferences.plist” in the “auto-install” or “.auto-install”
folder which resides in the same folder as the Tunnelblick.app.

Also adds a way to remove preferences, with a “remove” entry in
auto-install/preferences.plist, and always sets any preferences that
are not in a ‘remove’, ‘always-set’, or ‘set-only-if-not-present’ entry.