Fix UTF8 to bytes decoding problem
Fixed several places that used the length of a UTF8 string as the
number of bytes in the decoded UTF8, which is just plain wrong!

Caused config files with certain UTF8 sequences (in comments, for
example) to create a too-short file because the # of bytes output to
the file was wrong.