Fixes problems with renaming or duplicating a configuration
Fixes a problem that renaming a config caused -keychainHasUsername and
similar preferences to be set to FALSE in the new configuration even if
they weren’t set in the original configuration.

Fixes a problem that caused -keychainHasUsername and similar
preferences to be interpreted as TRUE even if they were FALSE!!!

Fixes a problem that caused a login window to appear when renaming or
duplicating a config if the -keychainHasUsername and similar
preferences did not reflect the true state of the Keychain.