Fix problem parsing a configuration with '-----BEGIN' but no '-----END'
To NEVER include anything that looks like certificate/key info, we
ignore everything after a line with ‘-----BEGIN’ until we see a line
with '-----END’, EVEN IF THE ‘-----BEGIN’ WAS INSIDE A COMMENT and
there is no ‘-----END’. That may mean we misread a file by ignoring
non-commented OpenVPN options. That’s a tradeoff that we accept.
However, we pretend the parse succeeded, so we can try to proceed.