Handle OpenVPN Challenge/Response static protocol.
Static protocol is activated when the configuration file (of the
client) contains the 'static-challenge' directive.
In that case, OpenVPN's management interface asks for the
challenge (the text of the challenge comes from the
configuration file - 'static-challenge' directive).

We build a popup window asking for the response
to that challenge, and we present the response to the server
(via the management interface).

[Documentation of static protocol](https://github.com/OpenVPN/openvpn/blob/master/doc/management-notes.txt#L1022)