Fix digital signature checking for Sierra
Checks digital signature as root when running from /Applications. This
is needed because Sierra changes permissions for some files that are
needed for the check so that they are only readable by root.

Also always does “deep” check. (Previously, a deep check was only
performed when running as the user.)