installer: properly detect failure to load tunnelblickd
The Tunnelblick installer did not properly detect the failure of
“launchctl load” to load tunnelblickd.

The launchctl man page says: “Due to bugs in the previous
implementation and long-standing client expectations around those bugs,
the load and unload subcommands will only return a non-zero exit code
due to improper usage.  Otherwise, zero is always returned.”

We now consider a “launchtl load” to have failed not only if a non-zero
exit code is returned, but also if either stdout or stderr from  the
command is not empty.