installer: force load of tunnelblickd even if it is disabled
The Tunnelblick installer did not force the loading of tunnelblickd if
loading tunnelblickd was disabled, which happens very rarely. (See
https://groups.google.com/d/msg/tunnelblick-discuss/fIHQiupUOVE/4CJeuenN
BgAJ.)

We now use the -w option of “launchctl load” to force the loading of
tunnelblickd.

The launchctl man page says: “-w Overrides the Disabled key and sets it
to false or true for the load and unload subcommands respectively.”