Pretend signature is valid on Mavericks & Higher
We sign this old version of Tunnelblick on 10.6.8 to get a version 1
signature that works on 10.5.0 - 10.8.5. But on 10.9 and higher, that
signature is invalid.

So someone running 10.9.0 or higher who downloads this version  or
updates to it, would get a warning from Tunnelblick that the program
has been tampered with.

We don’t want that to happen, and since such users are likely to run
this version only once before updating to a later version, we ignore
the signature if on 10.9 or higher.