Fix #429: Failure to enable 'Re-enable network access' after restart
Wait until all VPNs are disconnected if shutting down or restarting the
computer.

This is needed so the full disconnect of VPNs that disable network
access on disconnect works properly. Without the wait, some but not all
network services would be disabled, depending on timing (race condition
between disabling and the computer actually shutting down).