Implement .tblk Info.plist 'TBMinimumTunnelblickVersion' and 'TBMaximumTunnelblickVersion'
Refuses to install a .tblk if the version of Tunnelblick doesn’t
conform to the minimum/maximum version (if either or both were
specified) in the .tblk’s Info.plist (if it has one). Works for both
inner- and outer-level Info.plists. (That is, it works for .tblks
within a .tblk).