Fix new apparent public IP address showing as 0.0.0.0 when DNS fails
If DNS is not working but Internet access is OK, “Check if the apparent
public IP address changed after connecting” will try to check the IP
address by accessing tunnelblick.net using its IP address. When that’s
done successfully, an IP address of 0.0.0.0 is displayed instead of the
actual result. This fixes that problem.