Fix problem and speed up checking IP address
Fix the problem that App Transport Security blocked the http://
requests used to test if the IP address server was available by IP
address. Use an https: request, which causes an SSL failure because the
IP address is used, and then detect the SSL failure and consider that
proof that the server was contacted, so routing works OK.

Speeds things up (if there are failures) by eliminating the sleep
between requests and changing the timeout differently.