Refactor installer and implement EXPORT of Tunnelblick setup
 * Separate installer “OPERATIONS” (which there can be only one of per
invocation of installer) from “ACTIVITIES” (which can be done
optionally whenever installer is run). See defines.h for detailed
descriptions of the ACTIVITY bits and the OPERATION field.

 * Move much of the content of installer’s main() to subroutines.

 * Implement “Export Tunnelblick Setup” button on Utilities panel. It
creates a compressed .tblkSetup file on the user’s Desktop. The file
contains all Tunnelblick configurations, settings, and preferences for
all users. It *does*not* contain any items saved in the Keychain.