Implement user scripts before and after our standard scripts
Tunnelblick include three scripts for "Set username": up, down, and route-pre-down.

This allows the user to include "prefix" and/or "suffix" scripts for each one. These scripts are run at the start of and/or at the end of these scripts.

The user's scripts are provided with the arguments that OpenVPN provides to the up, down, or route-pre-down.

The up prefix and suffix scripts are run as root:wheel. The other scripts are run as root unless the OpenVPN configuration includes the 'user' or 'group' options.