Ignore Google Drive 'Icon?' files in a .tblk
Google Drive creates zero-length files named "Icon?" in each folder it synchronizes. Ignore them when installing a .tblk, so an install may be done by dragging/dropping from a Finder window showing the Google Drive.

Note: The "?" character in the filename is an ASCII LF character (at least sometimes). We ignore a file if it is empty and the filename begins with Icon and has one more character -- no matter what that character is.