Implement AppleScript commands for non-Tunnelblick configuration manipulation
Implements the following two new AppleScript commands:

"added or removed configurations", which should be used after adding and/or removing configurations so Tunnelblick can adjust its menu and the list of configurations in the "VPN Details" window.

"changed openvpn configuration file for NAME", where NAME (in double-quotes) is the name of a configuration file, which should be used after modifying the OpenVPN configuration file for a configuration so Tunnelblick can adjust settings which are affected by the OpenVPN configuration file.