Fix crash when certain network changes occur
Fixes a signal 11 when a network parameter changes to an empty value. That causes an empty line to appear in the log, which triggered an error accessing beyond the end of the string containing the line.

Also fixes an incorrect comment. The log now contains microseconds, not just whole seconds.