Fix delay in quitting Tunnelblick if user cancels install
If the user cancels the installation of Tunnelblick, tunnelblickd may not be set up, so if it is used when cleaning up before exiting, it won't respond and Tunnelblick will stall for 30 seconds waiting for it. This is annoying and might interfere with an attempt to install immediately afterward, so we skip cleaning up, which shouldn't be necessary anyway. (If we haven't fully launched, we haven't connected any VPNS or done anything else that needs to be saved or cleaned up, so this should be fine.)