Fix runTool to avoid multitasking/reentry problems
When runTool was executed on the main thread, it allowed events to be processed on the main thread while it was waiting for the tool to complete. This could cause errors/crashes.

We sleep instead of waiting, which forces the main thread to block. Because this is only done for quick "tools", that's acceptable.