Change "Disable IPv6" to "Disable IPv6 unless the VPN server is accessed using IPv6"
It makes no sense to disable IPv6 if the OpenVPN server is accessed via an IPv6 address. So change the checkbox title, and the meaning of the preference that it sets.

A warning is put in the Tunnelblick log if the checkbox is checked but IPv6 is not being disabled because the connection is to a VPN server with an IPv6 address.