Fix #539: Update to 3.7.9 while Tunnelblick is connected to VPN causes incorrect DNS settings
If a VPN is connected when Tunnelblick is updated, the OpenVPN process could crash and not restore the network setup.

This prevents that by making sure that even configurations that connect when the computer starts are disconnected if Tunnelblick is quitting because of an update. (Usually, they are not disconnected when Tunnelblick quits.)