Fix overwritten window contents when switching panels of VPN Details window
Similar to #537 by @m5o but changes only on Mojave and higher.

On macOS Mojave and higher, switching panels in the VPN Details window nowhappens instantly, which avoids the problem introduced in Mojave that caused macOS to ignore the "fade in" time.

On macOS El Capitan and lower, the panel switching behavior does not change: the panels are faded smoothly from one to another.