Create Tunnelblick tuntap installer .pkg
Create tunnelblick_tuntap_YYYYMMDD.pkg, an installer package modelled after the tuntaposx installer but which includes Tunnelblick's kexts.

If tap-notarized.kext or tun-notarized.kexts are present in third_party, they are included. Otherwise the newly built tap.kext and tun.kext are included.

A .tar.gz of the .pkg is placed in the tunnelblick/build/Release|Debug/ folder. The .pkg itself, and all supporting files and folders, are in the tunnelblick/build/Release|Debug/tuntap_pkg folder.