Added runningOn__Big__Sur__Successor__OrNewer() routine
Usually we add a runningOn...OrNewer() routine when the name of a new macOS version becomes known.

This time the new routine is being added now, before the new macOS version name is known, even though that means the routine's name will need to be changed when the name of the new macOS version is known.

It's being done now because we know there will be a problem if this routine is written to use the same pattern as the other runningOn...OrNewer() functions. See the new code's comments for details.

When the name of the OS version after Big Sur is announced, this routine should be renamed. (And tested thoroughly , too!)