Prepare to build for Apple Silicon
Prepare to build Tunnelblick and its third-party binaries for Apple Silicon as well as Intel.

PREPARATION ONLY:
 * DOES NOT BUILD FOR APPLE SILICON; and
 * DOES NOT BUILD ON APPLE SILICON!

Note: Sets Tunnelblick project in Xcode to build for "only active architecture" for both Debug and Release builds, so can build only on Intel. When building for and on Apple Silicon is working, set the Release build to build for both Intel and Apple Silicon.