Allow Xcode to do first build of third_party
Xcode 12 can't build some of the third_party code because it runs the script that builds that code as Apple Silicon code and some of the third_party code (lzo and LibreSSL, at least) won't build on an Apple Silicon host.

So we force the script that builds the third_party code to run in X86_64 mode (that is, using Rosetta 2).