Modify log change detection
The mechanism used to show the OpenVPN and Tunnelblick logs fails under high load because it queues each individual change to the log made by OpenVPN. OpenVPN can generate tens or hundreds of thousands of such changes per second, which can overload macOS's and Tunnelblick's ability to process them.

This modifies the mechanisms used to detect changes to the logs and to do throttling to prevent overloading the CPU.

1. Instead of using UKKQueue to react to any change to the log files, poll for changes to the log files every .25 seconds.

2. Instead of throttling to process only four changes to the log files at a time, discard any single change to the log file that is larger than 100,000 UTF-8 characters.

All loading of the log when it's display is empty (when the "Log" tab is switched to, for example) is limited to 4 * 100,000 UTF-8 characters.

 * The .25 can be overridden by the "connectionLogTickInterval" preference.

 * The 100,000 can be overridden by the "connectionLogEntrySizeLimit" preference.

 * The 4 can be overridden by the "connectionLogInitialLoadMultiplier" preference.