Fix #771
This changes the LaunchAtLogin.plist so the macOS Settings app shows the "Tunnelblick-LaunchAtLogin" login item with that name, instead of showing it with the name "Jonathan Bullard" (which is the name of the signing identity used to sign the executable).

It does not solve the problem that the program is listed under "Allow in Background" instead of under "Open at Login", which would be more accurate. That can apparently only be solved by changing the executable binary into a full app and including the app in Tunnelblick.app/Contents/LaunchAgents.