Fix crash when monitor network settings but not changing them
Fixes a bug introduced in 1ebf360e: don't monitor network settings if they are not being changed (i.e., if $foreign_option_1 is not defined or is empty).

Variables needed by process-network-changes were not set, causing it to crash when getting a substring from an uninitialized, and therefore zero-length string. Because that situation only happens when Tunnelblick isn't changing the network settings (because "$foreign_options_1" is empty), it isn't necessary (in most circumstances) to detect network changes, so we ignore the user's setting to detect them. (That setting was ignored in that situation until 1ebf360e, any would lead to a crash in 1ebf360e, so this "fix" shouldn't affect anyone.)