Fix #814: Unable to set permissions error when up command given an executable
This bug affects all OpenVPN options that accept a command as an argument (up, down, etc.).

Restriction on this solution: If an OpenVPN option that accepts a command as an argument uses a command that's an executable, the executable's path must be an absolute path. Otherwise the command will be considered to be invoking a script, and the script will be copied into the .tblk.