TBUpdater: Fix #837 "When restarting Mac, Tunnelblick always has a warning icon that it can't get an update"
If an update check fails because there is no Internet connection, complain only if:

 * This is a forced update check (i.e. the user clicked the "Check Now" button); or

 * There has never been a successful update check; or

 * It has been more than X seconds since the last successful check.

Where X is the "delayBeforeComplainingAboutFailedUpdateCheckBecauseInternetIsOffline" preference, which defaults to 7 days.

If no complaint is made, the (automatic) update check will be tried again in Y seconds, where Y is the "delayBeforeRetryingUpdateCheckBecauseInternetIsOffline" preference, which defaults to 60 seconds.

As before, for a forced check ("Check Now" button pushed), displaying errors in a pop-up window, otherwise only add a warning to the Tunnelblick menu and add a warning badge to the Tunnelblick icon.