Implement **only** OpenVPN's format option for static challenge
The earlier implementation of OpenVPN's "format" option for static challenge included other changes to integrate static challenge with username/password login screen, but this didn't work well because

(A) TOTP would sometimes expire before the user could enter the username, password, and TOTP, and

(B) the implementation removed the facility to use a script to set the response to the static challenge.

This commmit avoids those problems.