Ignore non-UTF-8 daemon requests without terminating the daemon
This fixes a problem reported by ZeroPath (even though it does not cause the local DoS that ZeroPath claims it does):

Non‑UTF8 request kills daemon
Denial of Service (DOS)

If a client sends a request whose payload is not valid UTF-8 (but otherwise satisfies the header and newline requirements), the code calls [NSString stringWithUTF8String:] and checks for NULL. On NULL it jumps to the 'done' label and proceeds to shut down the daemon (closing logs and returning). A local client that can connect to the listening socket can therefore terminate the root daemon with a single crafted request (simple local DoS).

NOTE: Non-UTF-8 requests do not cause a local DoS because the daemon will be restarted if/when a new request is made. However, there is no reason to terminate the daemon instead of just ignoring the request, so we just ignore the request.