Load full release notes display from appcast and then make changes
Instead of assembling the release notes HTML from already-translated pieces and the appcast's release notes which contain only the body:

Take the appcast's release notes as the HTML and then substitute the already-translated pieces for placeholder strings imbedded in the release notes.

This allows the window to be made taller by inserting empty space at the end of the <body> to compensate for the poor window sizing caused by using macOS build-in methods to calculate the window size.